Rename Site to AbstractSite
